third_party/blink/web_tests/external/wpt/service-workers/service-worker/worker-interception.https-expected.txt

This is a testharness.js-based test.
[FAIL] Verify a dedicated worker script request issued from a uncontrolled document is intercepted by worker's own service worker.
  promise_test: Unhandled rejection with value: object "Error: wait_for_state needs a ServiceWorker object to be passed."
[FAIL] Requests on a dedicated worker nested in a shared worker and controlled by a service worker
  assert_equals: expected "This load was successfully intercepted." but got "Unexpected error! Worker is not defined"
Harness: the test ran to completion.
